Work on updating tests for handling funds
